// javascript/packages/nextjs/src/internal/proto/google/protobuf/struct.ts
export enum NullValue {
  /** NULL_VALUE - Null value. */
  NULL_VALUE = 0,
  UNRECOGNIZED = -1,
}